Extract error types into own module
